requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.6
Django==5.2.4
django-allauth==65.10.0
django-filter==25.1
django-modeltranslation==0.19.16
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
openapi-codec==1.3.2
packaging==25.0
phonenumbers==9.0.10
phonenumberslite==9.0.10
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
simplejson==3.20.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0